pyproject.toml
[tool.poetry]
name = "project-snow"
version = "0.1.0"
description = ""
authors = ["wookiee89 <jay.r.jain@gmail.com>"]
readme = "README.md"
packages = [{include = "project_snow"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.101.0"
pydantic = {extras = ["dotenv"], version = "^2.1.1"}
uvicorn = "^0.23.2"
SQLAlchemy = {extras = ["asyncio"], version = "^2.0.19"}
asyncpg = "^0.28.0"
pydantic-settings = "^2.0.2"
auth0-python = "^4.4.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.1"
httpx = "^0.24.1"
pylint = "^2.15.9"
ruff = "^0.0.284"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
# Enable the pycodestyle (`E`) and Pyflakes (`F`) rules by default.
# Unlike Flake8, Ruff doesn't enable pycodestyle warnings (`W`) or
# McCabe complexity (`C901`) by default.

line-length = 100  # defaults to 88 like black


select = [
  "E",   # pycodestyle
  "F",   # pyflakes
  "UP",  # pyupgrade,
  "I",   # isort
]
ignore = []

# Allow autofix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]
per-file-ignores = {}

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"